version 0.5

+++ src/cli/Evaluator.java
@@ -6,7 +6,7 @@ public class Evaluator {
     public static void main(String[] args) throws Exception {
         EvaluatorLib evl = new EvaluatorLib();
         try {
-            int res = evl.evaluateExpression(args[0]);
+            double res = evl.evaluateExpression(args[0]);
             System.out.println(res);
         }
         catch (Exception e){
